Added a `resolve()` function to resolve alerts.
There is also a new `pd-resolve` command-line tool that can resolve incidents
for you. Refactored the module slightly so that the http request is in a utility
function.

Added the `nyc` code coverage tool to the default test suite.
Added tests until we're at 100% code coverage.
Documented the new feature.

Added a license file to make it clear what the module license is.